+++ components/stratum/include/mining.h
@@ -18,7 +18,7 @@ typedef struct
     uint8_t midstate1[32];
     uint8_t midstate2[32];
     uint8_t midstate3[32];
-    uint32_t pool_diff;
+    double pool_diff;
     char *jobid;
     char *extranonce2;
 } bm_job;
@@ -30,7 +30,7 @@ void calculate_coinbase_tx_hash(const char *coinbase_1, const char *coinbase_2,
 
 void calculate_merkle_root_hash(const uint8_t coinbase_tx_hash[32], const uint8_t merkle_branches[][32], const int num_merkle_branches, uint8_t dest[32]);
 
-void construct_bm_job(mining_notify *params, const uint8_t merkle_root[32], const uint32_t version_mask, const uint32_t difficulty, bm_job* new_job);
+void construct_bm_job(mining_notify *params, const uint8_t merkle_root[32], const uint32_t version_mask, const double difficulty, bm_job* new_job);
 
 double test_nonce_value(const bm_job *job, const uint32_t nonce, const uint32_t rolled_version);
